Project/auth_service/requirements.txt
Django
django-filter
django-restframework
djangorestframework
djangorestframework-simplejwt
drf-yasg
markdown
psycopg
psycopg2-binary
requests